Modify move-to-workspace-up script to allow for smooth transitions (#1)
The previous implementation moved the current window to the last
workspace, then moved that workspace to be the first. This worked but
caused the animations to break.

The change in this commit moves the window to the last workspace in the
background (Without focusing it), moves that workspace to be the first
and then focuses it, making it look like any other Niri transition.